Initial "as is" import
The code compiles with make. VS and Eclipse configuration files are included,
but they probably need an update.
This version of Metroscope uses Craftags, the ancestors of Chilitags,
which are compatible (the tags look the same).
This release includes the quick and dirty code needed to generate the "planar
calibration", i.e. the mapping of points in the image camera to points in a
calibrated surface, and of the points in this real world surface to points in
projected image. This subsystem can be simplified a lot by using tags instead
a chessboard.
"metroscope" will run, but to do anything, it will need configuration files.
To do anything useful, it will need calibration files. None are included in
this repository (for now).
This code comes from 2 years of development, with almost no documentation,
good luck.